pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "scikit-rec"
dynamic = ["version"]
description = "A scikit-style recommender systems library"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.13.3,<2.0.0",
    "pandas>=1.5.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "scikit-learn>=1.4.2,<2.0.0",
    "xgboost>=2.1.0,<4.0.0",
    "lightgbm>=4.6.0",
    "pyarrow>=10.0.0,<20.0.0",
    "optuna>=3.0.0,<5.0.0",
]

[project.optional-dependencies]
torch = [
    "torch>=2.2.0,<3.0.0",
]
aws = [
    "boto3>=1.26.27,<3.0.0",
]
explain = [
    "shap>=0.44.0,<1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "moto[s3]>=4.0",
    "pre-commit>=3.0",
]

[project.urls]
Homepage = "https://github.com/intuit/scikit-rec"
Repository = "https://github.com/intuit/scikit-rec"
Issues = "https://github.com/intuit/scikit-rec/issues"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
include = ["skrec*"]

# --- ruff (lint + format) ---
[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle errors, pyflakes, isort

[tool.ruff.format]
quote-style = "double"

# --- mypy ---
[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

# --- pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

# --- coverage ---
[tool.coverage.run]
source = ["skrec"]

[tool.coverage.report]
fail_under = 80
